+++ commands/exit.gd
@@ -1,0 +1,4 @@
+extends Node
+
+static func execute(who, _anim):
+	who.get_tree().quit()