+++ commands/fullscreen.gd
@@ -1,0 +1,5 @@
+extends Node
+
+static func execute(_who, _anim):
+	OS.window_borderless = ! OS.window_borderless
+	OS.window_fullscreen = ! OS.window_fullscreen