+++ commands/set_name.gd
@@ -1,0 +1,4 @@
+extends Node
+
+static func execute(_who, name):
+	Data.set_char_value("name", name)