+++ singletons/data.gd
@@ -25,7 +25,12 @@ extends Node
 var abilities: Dictionary = {}
 var items: Dictionary = {}
 var mobiles: Dictionary = {}
-var config: Dictionary = {"invert_x":1, "invert_y":1, "fullscreen":false}
+var config: Dictionary = {
+		"invert_x":1, 
+		"invert_y":-1, 
+		"fullscreen":false,
+		"last_character":"new_character",
+			}
 var char_data: Dictionary = {"name":"New Character", "equipment":["base_human_body"], "chat_color":"lime"}
 
 func _init() -> void:
@@ -48,14 +53,16 @@ func init_lists() -> void:
 	abilities = load_from_file("abilities")
 	items = load_from_file("items")
 	mobiles = load_from_file("mobiles")
-	config = load_from_file("config", "user://", "")
+	load_config()
+	load_char_data(get_config_value("last_character"))
 
 func load_from_file(file_name: String, path: String = "res://data/json/", extension: String = ".json") -> Dictionary:
 	var f = File.new()
 	var list = {}
-	f.open(path+file_name+extension, File.READ)
-	list = JSON.parse(f.get_as_text()).result
-	f.close()
+	if f.file_exists(path + file_name + extension):
+		f.open(path+file_name+extension, File.READ)
+		list = JSON.parse(f.get_as_text()).result
+		f.close()
 	return list
 
 func get_char_data() -> Dictionary:
@@ -68,7 +75,19 @@ func get_char_value(key: String):
 		return "Char data does not contain " + key
 	
 func load_char_data(c: String) -> void:
-	char_data = load_from_file(c, "user://saves/", "")
+	var f = File.new()
+	if f.file_exists("user://saves/"+c):
+		char_data = load_from_file(c, "user://saves/", "")
+	else:
+		print(c + " not found in the saves folder.")
+		
+func load_config() -> void:
+	var f = File.new()
+	if f.file_exists("user://config"):
+		var inc = load_from_file("config", "user://", "")
+		for i in config:
+			if inc.has(i):
+				config[i] = inc[i]
 	
 func get_config_value(key: String): 
 	if config.has(key):
@@ -81,6 +100,15 @@ func get_snake_case(s: String) -> String:
 	s = s.to_lower()
 	s = s.replace(" ", "_")
 	return s
+	
+func set_char_value(key: String, value) -> void:
+	if char_data.has(key):
+		char_data[key] = value
+	Events.emit_signal("char_data_changed", key, value)
+		
+func set_config_value(key: String, value) -> void:
+	if config.has(key):
+		config[key] = value
 
 func save_char() -> void:
 	var f = File.new()
@@ -95,3 +123,8 @@ func save_config() -> void:
 	f.open("user://"+"config", File.WRITE)
 	f.store_string(JSON.print(d))
 	f.close()
+
+func _exit_tree():
+	set_config_value("last_character", get_snake_case(get_char_value("name")))
+	save_config()
+	save_char()