+++ scenes/scene_manager.gd
@@ -7,12 +7,13 @@ extends Spatial
 	Setup : 
 """
 
-export(String) var start_scene = "main"
+export(String) var start_scene = "char_creation"
 
 var current_scene = null
 
 func _ready():
 	switch_scene(start_scene)
+	Events.connect("scene_change_request", self, "switch_scene")
 
 func switch_scene(scene:String ) -> void:
 	if is_instance_valid(current_scene):