+++ singletons/events.gd
@@ -13,3 +13,4 @@ signal chat_message_received
 signal ui_opened
 signal ui_closed
 signal char_data_changed
+signal scene_change_request