+++ mobiles/armatures/sensors.gd
@@ -1,7 +1,19 @@
 extends Spatial
 
+onready var over_ledge = $Overledge
+onready var under_ledge = $Overledge/Underledge
+onready var interact = $Interact
+
 func get_ledge():
-	if ! $Overledge.is_colliding() and $Overledge/Underledge.is_colliding():
-		return $Overledge/Underledge.get_collision_normal()
+	if ! over_ledge.is_colliding() and under_ledge.is_colliding():
+		return under_ledge.get_collision_normal()
 	else:
 		return Vector3.ZERO
+		
+func get_interact_target():
+	if interact.is_colliding() and interact.get_collider().has_method("interact"):
+		$Interact/Label.text = interact.get_collider().name
+		return interact.get_collider()
+	else:
+		$Interact/Label.text = ""
+		return null