+++ objects/interactable/debug_cube.gd
@@ -3,4 +3,3 @@ extends StaticBody
 func interact(target) -> void:
 	if target.is_in_group("actors"):
 		target.global_transform.origin = $Position3D.global_transform.origin
-	print("interact")