+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [Unreleased]
 
+[Removed] i_controller
 [Added] controlled camera can now zoom in and out with mouse wheel
 [Added] a raycast was added to armature/sensors for detecting interactive objects
 [Changed] Data.load_char_save() now handles get_snake_case(), so all saves should be converted to snake_case and should not be considered case-sensitive